recommendation-service/requirements.txt
flask==3.1.1
python-dotenv==1.1.0
pymysql==1.1.1
scikit-learn==1.7.0
flask-cors==6.0.1